+++ packages/core/src/lib/types.ts
@@ -3,6 +3,10 @@ import { Signal, WritableSignal } from '@angular/core';
 import { AsyncQueryMarker, AsyncQuerySignal } from './markers/async-query';
 import { AsyncSourceMarker, AsyncSourceSignal } from './markers/async-source';
 import { AsyncStreamMarker, AsyncStreamSignal } from './markers/async-stream';
+import {
+  EntityCollectionMarker,
+  EntityCollectionSignal,
+} from './markers/entity-collection';
 import { FormMarker, FormSignal } from './markers/form';
 import { StatusMarker, StatusSignal } from './markers/status';
 import { StoredMarker, StoredSignal } from './markers/stored';
@@ -103,7 +107,9 @@ export interface NodeAccessor<T> {
 // Default TreeNode maps known keys to either EntitySignal, StatusSignal, StoredSignal, FormSignal,
 // or CallableWritableSignal and still allows dynamic string indexing at runtime.
 export type TreeNode<T> = {
-  [K in keyof T]: T[K] extends EntityMapMarker<infer E, infer Key>
+  [K in keyof T]: T[K] extends EntityCollectionMarker<infer CE, infer CK>
+    ? EntityCollectionSignal<CE, CK>
+    : T[K] extends EntityMapMarker<infer E, infer Key>
     ? EntitySignal<E, Key>
     : T[K] extends StatusMarker<infer Err>
     ? StatusSignal<Err>
@@ -776,7 +782,9 @@ export type IsEntityMap<T> = T extends EntityMapMarker<
  * the full deep inference.
  */
 export type DeepEntityAwareTreeNode<T> = {
-  [K in keyof T]: T[K] extends EntityMapMarker<infer E, infer Key>
+  [K in keyof T]: T[K] extends EntityCollectionMarker<infer CE, infer CK>
+    ? EntityCollectionSignal<CE, CK>
+    : T[K] extends EntityMapMarker<infer E, infer Key>
     ? EntitySignal<E, Key>
     : T[K] extends StatusMarker<infer Err>
     ? StatusSignal<Err>
@@ -803,7 +811,9 @@ export type DeepEntityAwareTreeNode<T> = {
  * `TypedSignalTree<T>` (see below) or use `DeepEntityAwareTreeNode`.
  */
 export type EntityAwareTreeNode<T> = {
-  [K in keyof T]: T[K] extends EntityMapMarker<infer E, infer Key>
+  [K in keyof T]: T[K] extends EntityCollectionMarker<infer CE, infer CK>
+    ? EntityCollectionSignal<CE, CK>
+    : T[K] extends EntityMapMarker<infer E, infer Key>
     ? EntitySignal<E, Key>
     : T[K] extends StatusMarker<infer Err>
     ? StatusSignal<Err>